Increased maximum number of builtin functions
Number of builtins surpassed 256 so needed to upgrade from
a uint8_t to a uint16_t

+++ src/include/execution/ast/builtins.h
@@ -319,7 +319,7 @@ namespace terrier::execution::ast {
 /**
  * An enumeration of all TPL builtin functions.
  */
-enum class Builtin : uint8_t {
+enum class Builtin : uint16_t {
 #define ENTRY(Name, ...) Name,
   BUILTINS_LIST(ENTRY)
 #undef ENTRY
@@ -350,7 +350,7 @@ class Builtins {
   /**
    * @return The name of the function associated with the given builtin enumeration.
    */
-  static const char *GetFunctionName(Builtin builtin) { return builtin_function_names[static_cast<uint8_t>(builtin)]; }
+  static const char *GetFunctionName(Builtin builtin) { return builtin_function_names[static_cast<uint16_t>(builtin)]; }
 
  private:
   static const char *builtin_function_names[];
